chore: move ws-rpc and socket-io to ws/ directory
