+++ src/api/init.ts
@@ -26,8 +26,8 @@ import { createRosettaBlockRouter } from './routes/rosetta/block';
 import { createRosettaAccountRouter } from './routes/rosetta/account';
 import { createRosettaConstructionRouter } from './routes/rosetta/construction';
 import { isProdEnv, logError, logger, LogLevel, waiter } from '../helpers';
-import { createWsRpcRouter } from './routes/ws-rpc';
-import { createSocketIORouter } from './routes/socket-io';
+import { createWsRpcRouter } from './routes/ws/ws-rpc';
+import { createSocketIORouter } from './routes/ws/socket-io';
 import { createBurnchainRouter } from './routes/burnchain';
 import { createBnsNamespacesRouter } from './routes/bns/namespaces';
 import { createBnsPriceRouter } from './routes/bns/pricing';
